Adjust the deadline reminders.

+++ app/Console/Commands/DeadlineReminderCommand.php
@@ -7,6 +7,7 @@ use Illuminate\Support\Facades\Mail;
 use App\Mail\StudentRemovedFromStudyAdminEmail;
 use App\Mail\UserRunningOutOfTimeEmail;
 use App\Models\User;
+use Carbon\Carbon;
 
 class DeadlineReminderCommand extends Command
 {
@@ -51,17 +52,24 @@ class DeadlineReminderCommand extends Command
         if (!empty($users)) {
             foreach ($users as $user) {
                 foreach ($user->moduleProgress as $currentModule) {
-                    $dayCheck = $currentModule->created_at->diffInDays($currentModule->expiration_date);
-                    if ($dayCheck === 2 || $dayCheck === 1) {
-                        if ($currentModule->current_page !== $currentModule->max_page) {
-                            // send out the email.
-                            Mail::to($user->email)->cc(env('RECEIVE_EMAIL'))->send(new UserRunningOutOfTimeEmail($user, $currentModule->current_module));
+                    // we want to check today's date vs the expiration date of the module.
+                    $dayCheck = Carbon::now()->diffInDays($currentModule->expiration_date);
+                    if ($user->getUserGroup->user_group !== 'comparison') {
+                        if ($dayCheck === 2 || $dayCheck === 1) {
+                            if ($currentModule->current_page !== $currentModule->max_page) {
+                                // send out the email.
+                                Mail::to($user->email)->cc(env('RECEIVE_EMAIL'))->send(new UserRunningOutOfTimeEmail($user, $currentModule->current_module));
+                            }
                         }
-                    }
-                    if ($dayCheck === 0) {
-                        if ($currentModule->current_page === 0 && $currentModule->max_page === 0) {
-                            $user->participant()->delete();
-                            Mail::to(env('RECEIVE_EMAIL'))->send(new StudentRemovedFromStudyAdminEmail($user, $currentModule->current_module));
+                        if ($dayCheck === 0) {
+                            if ($currentModule->current_page === 0 && $currentModule->max_page === 0) {
+                                $user->participant()->delete();
+                                Mail::to(env('RECEIVE_EMAIL'))->send(new StudentRemovedFromStudyAdminEmail($user, $currentModule->current_module));
+                            }
+                        }
+                    } else {
+                        if ($dayCheck === 10 || $daCheck === 3)  {
+                            Mail::to($user->email)->cc(env('RECEIVE_EMAIL'))->send(new UserRunningOutOfTimeEmail($user, $currentModule->current_module));
                         }
                     }
                 }
